requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
gunicorn==20.1.0
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
PyJWT==2.6.0
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
typing_extensions==4.4.0
tzdata==2022.7
virtualenv==20.17.1